Split review screen components
