+++ apps/android/feature/review/src/main/java/com/flashcardsopensourceapp/feature/review/ReviewSharedComponents.kt
@@ -1,0 +1,27 @@
+package com.flashcardsopensourceapp.feature.review
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Card
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+internal val reviewMetadataIconSize = 18.dp
+
+@Composable
+internal fun LoadingReviewState() {
+    Card(modifier = Modifier.fillMaxWidth()) {
+        Box(
+            contentAlignment = Alignment.Center,
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(32.dp)
+        ) {
+            CircularProgressIndicator()
+        }
+    }
+}
